+++ src/main/java/com/cburch/logisim/gui/menu/MenuFile.java
@@ -34,6 +34,7 @@ class MenuFile extends Menu implements ActionListener {
   private final JMenuItem save = new JMenuItem();
   private final JMenuItem saveAs = new JMenuItem();
   private final JMenuItem exportProj = new JMenuItem();
+  private final JMenuItem extractRunProj = new JMenuItem();
   private final MenuItemImpl print = new MenuItemImpl(this, LogisimMenuBar.PRINT);
   private final MenuItemImpl exportImage = new MenuItemImpl(this, LogisimMenuBar.EXPORT_IMAGE);
   private final JMenuItem prefs = new JMenuItem();
@@ -63,6 +64,8 @@ class MenuFile extends Menu implements ActionListener {
     add(close);
     add(save);
     add(saveAs);
+    addSeparator();
+    add(extractRunProj);
     add(exportProj);
     addSeparator();
     add(exportImage);
@@ -85,12 +88,14 @@ class MenuFile extends Menu implements ActionListener {
       save.setEnabled(false);
       saveAs.setEnabled(false);
       exportProj.setEnabled(false);
+      extractRunProj.setEnabled(false);
     } else {
       merge.addActionListener(this);
       close.addActionListener(this);
       save.addActionListener(this);
       saveAs.addActionListener(this);
       exportProj.addActionListener(this);
+      extractRunProj.addActionListener(this);
     }
     menubar.registerItem(LogisimMenuBar.EXPORT_IMAGE, exportImage);
     menubar.registerItem(LogisimMenuBar.PRINT, print);
@@ -163,6 +168,8 @@ class MenuFile extends Menu implements ActionListener {
         ProjectActions.doSaveAs(proj);
       } else if (src == exportProj) {
         ProjectActions.doExportProject(proj);
+      } else if (src == extractRunProj) {
+        ProjectActions.doExtractAndRunProject(proj);
       }
     }
   }
@@ -183,6 +190,7 @@ class MenuFile extends Menu implements ActionListener {
     save.setText(S.get("fileSaveItem"));
     saveAs.setText(S.get("fileSaveAsItem"));
     exportProj.setText(S.get("fileExportProject"));
+    extractRunProj.setText(S.get("fileExtractRunProject"));
     exportImage.setText(S.get("fileExportImageItem"));
     print.setText(S.get("filePrintItem"));
     prefs.setText(S.get("filePreferencesItem"));