+++ CHANGES.md
@@ -6,6 +6,7 @@
   * Fixed bug preventing TTL 7442, 7443 and 7444 from being placed on the circuit canvas.
   * Sub-circuit can now be deleted with `DELETE` key, along with `BACKSPACE` used so far.
   * Fixed `Simulate` -> `Timing Diagram` not opening when using "Nimbus" look and feel.
+  * New take on project export/import a zip-file is generated which can include a user provided "README.md".
 
 * v3.7.2 (2021-11-09)
   * Fixed Preferences/Window "Reset window layout to defaults" not doing much.
@@ -39,6 +40,7 @@
   * Added a setting to select lower- or upper-case VHDL keywords.
   * Added project export feature.
   * Cleaned-up the written .circ file.
+  * Cleaned-up the library tree of loaded projects.
 
 * v3.6.1 (2021-09-27)
   * Fixed bug in LED-array